Re-organize types
